+++ modules/localgov_review_date/README.md
@@ -1,0 +1,3 @@
+# LocalGov Drupal Review Date
+
+Provides a review dates entity and widget, along with views for managing content reviews.
