+++ modules/localgov_review_date/localgov_review_date.links.task.yml
@@ -1,0 +1,5 @@
+localgov_review_date.needs_review:
+  title: 'Needs review'
+  route_name: view.localgov_needs_review.needs_review
+  parent_id: system.admin_content
+  weight: 2
