+++ modules/localgov_review_date/localgov_review_date.module
@@ -1,0 +1,64 @@
+<?php
+
+/**
+ * @file
+ * Localgov Review Date module hooks.
+ */
+
+use Drupal\Core\Entity\EntityTypeInterface;
+use Drupal\Core\Field\BaseFieldDefinition;
+
+/**
+ * Implements hook_entity_bundle_field_info().
+ */
+function localgov_review_date_entity_bundle_field_info(EntityTypeInterface $entity_type, $bundle, array $base_field_definitions) {
+
+  // Add the review status field to nodes configured for scheduled transitions.
+  if ($entity_type->id() === 'node') {
+
+    $scheduled_config = \Drupal::service('config.factory')->get('scheduled_transitions.settings');
+    if (is_null($scheduled_config)) {
+      return;
+    }
+
+    $scheduled_bundles = $scheduled_config->get('bundles');
+    if (is_null($scheduled_bundles)) {
+      return;
+    }
+
+    foreach ($scheduled_bundles as $scheduled_bundle) {
+      if ($scheduled_bundle['entity_type'] == $entity_type->id() && $scheduled_bundle['bundle'] == $bundle) {
+        $fields['localgov_review_date'] = BaseFieldDefinition::create('review_date')
+          ->setLabel(t('Review date'))
+          ->setDisplayOptions('form', [
+            'type' => 'review_status',
+            'weight' => -5,
+          ])
+          ->setDisplayConfigurable('form', TRUE)
+          ->setComputed(TRUE);
+
+        return $fields;
+      }
+    }
+  }
+}
+
+/**
+ * Implements hook_form_BASE_FORM_ID_alter().
+ *
+ * Add submit handler to the scheduled transitions settings form.
+ */
+function localgov_review_date_form_scheduled_transitions_settings_form_alter(&$form, \Drupal\Core\Form\FormStateInterface $form_state, $form_id) {
+  $form['#submit'][] = 'localgov_review_date_form_scheduled_transitions_settings_submit';
+}
+
+/**
+ * Clear caches after submitting the scheduled transitions settings form.
+ */
+function localgov_review_date_form_scheduled_transitions_settings_submit(array $form, \Drupal\Core\Form\FormStateInterface $form_state) {
+
+  // The review date field is only added to nodes that have scheduled
+  // transitions enabled. It's necessary to rebuild caches after making changes
+  // to scheduled transitions to ensure the field is added.
+  drupal_flush_all_caches();
+}
