--- modules/localgov_review_status/README.md
@@ -1,3 +1,0 @@
-# LocalGov Drupal Review Status
-
-Provides a review dates entity and widget, along with views for managing content reviews.
