--- modules/localgov_review_status/localgov_review_status.libraries.yml
@@ -1,6 +1,0 @@
-localgov_review_status.review_status:
-  js:
-    js/review-status.js: {}
-  dependencies:
-    - core/drupal.entity-form
-    - core/drupalSettings
