--- modules/localgov_review_status/localgov_review_status.links.task.yml
@@ -1,5 +1,0 @@
-localgov_review_status.needs_review:
-  title: 'Needs review'
-  route_name: view.localgov_needs_review.needs_review
-  parent_id: system.admin_content
-  weight: 2
