--- modules/localgov_review_status/localgov_review_status.permissions.yml
@@ -1,3 +1,0 @@
-administer localgov_review_status:
-  title: 'Administer LocalGov Review Status'
-  description: 'Manage LocalGov review status settings.'
